Custom_New 更新FileUpload，删除OSS对象存储

+++ src/main/java/controllers/AppErrorLogController.java
@@ -1,13 +1,16 @@
 package controllers;
 
 import annotation.Controller;
+import base.Constant;
 import base.controller.BaseApiController;
 import com.jfinal.core.ActionKey;
+import com.jfinal.kit.Prop;
 import com.jfinal.upload.UploadFile;
 import models.member.MemberLogin;
 import models.sys.Applog;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import services.FileService;
 import utils.AlyFileUpload;
 import utils.DateUtil;
 
@@ -27,34 +30,41 @@ public class AppErrorLogController extends BaseApiController {
      */
     @ActionKey("/appErrorLog")
     public void appErrorLog() {
-        Applog applog = new Applog();
-        MemberLogin memberLogin = null;
-        if (getLoginMember() != null) {
+        Applog applog =new Applog();
+        MemberLogin memberLogin=null;
+        if(getLoginMember() != null){
             memberLogin = MemberLogin.dao.findById(getLoginMember().getId());
         }
-        applog.setAppType(getAppType());
+        applog.setAppType( getAppType());
         applog.setDeviceType(getDeviceType());
         UploadFile uf = getFile();
         File file = uf.getFile();
+        String uploadDir = Constant.properties.getProperty("apperrorlog.dir", "");
+        String uploadHost = Constant.properties.getProperty("apperrorlog.host", "");
+        uploadDir = FileService.getInstance().mkdir(uploadDir);
         String fileType = file.getName();
         Integer versionCode = 0;
-        if (fileType.indexOf("C") != -1) {
-            String[] strings = fileType.split("C");
+        if(fileType.indexOf("C")!=-1){
+            String[]  strings =  fileType.split("C");
             fileType = strings[1];
             versionCode = Integer.parseInt(strings[0]);
             applog.setVersionCode(versionCode);
         }
+
         String dateName = DateUtil.stampToDate(fileType.substring(0, fileType.lastIndexOf(".")));
+        String filePath= uploadDir.substring(uploadDir.indexOf("appLog"), uploadDir.length());
+        File t = new File(uploadDir +(memberLogin!=null ?memberLogin.getUserName():"N")+"-"+ dateName.replace(":","--")+".txt");
         try {
-            String fileName = (memberLogin != null ? memberLogin.getUserName() : "N") + "-" + dateName.replace(":", "--") + ".txt";
-            String resultPath = AlyFileUpload.getIntance().uploadByInputStream(file, fileName);
-            applog.setLogPath(resultPath);
-            file.delete();
-            applog.setCreateTime(new Date());
-            if (applog.save()) {
-                renderAjaxSuccess(resultPath);
-            } else {
-                throw new IOException();
+            if (t.createNewFile()) {
+                FileService.getInstance().fileChannelCopy(file, t);
+                file.delete();
+                applog.setLogPath(uploadHost +filePath+(memberLogin!=null ?memberLogin.getUserName() : "N") + "-" + dateName.replace(":", "--") + ".txt");
+                applog.setCreateTime(new Date());
+                if(applog.save()){
+                    renderAjaxSuccess(uploadHost+filePath + (memberLogin != null ? memberLogin.getUserName() : "N") + "-" + dateName.replace(":", "--") + ".txt");
+                }else {
+                    throw new IOException();
+                }
             }
         } catch (IOException e) {
             e.printStackTrace();
